% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{autoDetectImageName}
\alias{autoDetectImageName}
\title{Auto-detect image name from current directory}
\usage{
autoDetectImageName(image_name = NULL)
}
\arguments{
\item{image_name}{Existing image name (if provided)}
}
\value{
Image name (lowercased directory basename)
}
\description{
Auto-detect image name from current directory
}
\keyword{internal}
