% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{buildDockerLabels}
\alias{buildDockerLabels}
\title{Build Docker labels for OmopStudyBuilder containers}
\usage{
buildDockerLabels(image_name, mode)
}
\arguments{
\item{image_name}{Docker image name}

\item{mode}{Container mode ("rstudio" or "run")}
}
\value{
Character vector of Docker label arguments
}
\description{
Build Docker labels for OmopStudyBuilder containers
}
\keyword{internal}
