% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{checkDocker}
\alias{checkDocker}
\title{Check if Docker is available}
\usage{
checkDocker()
}
\value{
Logical. TRUE if Docker is installed and reachable, FALSE otherwise
}
\description{
Check if Docker is available
}
