% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{checkRepoAvailable}
\alias{checkRepoAvailable}
\title{Check if repository name is available}
\usage{
checkRepoAvailable(owner, repo)
}
\description{
Check if repository name is available
}
\keyword{internal}
