% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{createCodeReviewIssue}
\alias{createCodeReviewIssue}
\title{Create code review checklist issue on GitHub}
\usage{
createCodeReviewIssue(owner, repo)
}
\description{
Create code review checklist issue on GitHub
}
\keyword{internal}
