% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{createGitHubRepo}
\alias{createGitHubRepo}
\title{Create GitHub repository via API}
\usage{
createGitHubRepo(repository, organisation, private, description)
}
\description{
Create GitHub repository via API
}
\keyword{internal}
