% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{createPhenotyperSettingsIssue}
\alias{createPhenotyperSettingsIssue}
\title{Open a GitHub issue listing PhenotypeR diagnostics settings to review}
\usage{
createPhenotyperSettingsIssue(owner, repository)
}
\description{
Open a GitHub issue listing PhenotypeR diagnostics settings to review
}
\keyword{internal}
