% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{createReviewChecklistIssue}
\alias{createReviewChecklistIssue}
\title{Open a GitHub issue with a code review checklist}
\usage{
createReviewChecklistIssue(owner, repository)
}
\description{
Open a GitHub issue with a code review checklist
}
\keyword{internal}
