% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{createStudyGitIgnore}
\alias{createStudyGitIgnore}
\title{Create .gitignore for OMOP studies}
\usage{
createStudyGitIgnore(directory)
}
\description{
Create .gitignore for OMOP studies
}
\keyword{internal}
