% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{detectRVersion}
\alias{detectRVersion}
\title{Detect and sync R version for Docker}
\usage{
detectRVersion(dockerfile_dest, study_path = NULL)
}
\description{
Detect and sync R version for Docker
}
\keyword{internal}
