% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{dockerExec}
\alias{dockerExec}
\title{Execute a Docker command}
\usage{
dockerExec(args, error_message = "Docker command failed")
}
\arguments{
\item{args}{Character vector of Docker command arguments}

\item{error_message}{Custom error message if command fails}
}
\value{
Command output as character vector (returned invisibly to avoid console clutter)
}
\description{
Execute a Docker command
}
\keyword{internal}
