% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{ensureDocker}
\alias{ensureDocker}
\title{Check docker binary and (optionally) daemon (internal)}
\usage{
ensureDocker(check_daemon = FALSE)
}
\description{
Check docker binary and (optionally) daemon (internal)
}
\keyword{internal}
