% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{ensureGit}
\alias{ensureGit}
\title{Check if Git is installed}
\usage{
ensureGit()
}
\value{
TRUE if Git is available, throws error otherwise
}
\description{
Check if Git is installed
}
\keyword{internal}
