% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{findAvailablePort}
\alias{findAvailablePort}
\title{Find next available port starting from a given port}
\usage{
findAvailablePort(start_port = 8787, max_tries = 10)
}
\arguments{
\item{start_port}{Starting port number to check}

\item{max_tries}{Maximum number of ports to try}
}
\value{
Available port number
}
\description{
Find next available port starting from a given port
}
\keyword{internal}
