% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{formatHostForUrl}
\alias{formatHostForUrl}
\title{Format a host for use in an http URL}
\usage{
formatHostForUrl(host)
}
\arguments{
\item{host}{Host string}
}
\value{
Host formatted for URL (adds brackets for IPv6)
}
\description{
Format a host for use in an http URL
}
\keyword{internal}
