% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{getCodeReviewChecklistBody}
\alias{getCodeReviewChecklistBody}
\title{Get code review checklist template}
\usage{
getCodeReviewChecklistBody()
}
\description{
Get code review checklist template
}
\keyword{internal}
