% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{gitCommit}
\alias{gitCommit}
\title{Create git commit}
\usage{
gitCommit(
  directory,
  message,
  allow_nothing_to_commit = FALSE,
  error_message = "Failed to commit"
)
}
\description{
Create git commit
}
\keyword{internal}
