% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{imageExists}
\alias{imageExists}
\title{Check if a Docker image exists locally}
\usage{
imageExists(image_name)
}
\arguments{
\item{image_name}{Name of Docker image}
}
\value{
TRUE if the image exists, FALSE otherwise
}
\description{
Check if a Docker image exists locally
}
\keyword{internal}
