% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{imageHasRStudioServer}
\alias{imageHasRStudioServer}
\title{Check if a Docker image contains RStudio Server}
\usage{
imageHasRStudioServer(image_name)
}
\arguments{
\item{image_name}{Name of Docker image}
}
\value{
TRUE if rserver is present, FALSE otherwise
}
\description{
Check if a Docker image contains RStudio Server
}
\keyword{internal}
