% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{initRenv}
\alias{initRenv}
\title{Initialize renv in a study directory}
\usage{
initRenv(study_path)
}
\arguments{
\item{study_path}{Path to the study directory}
}
\value{
Invisibly returns TRUE
}
\description{
Initialize renv in a study directory
}
