% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{prepareDockerfile}
\alias{prepareDockerfile}
\title{Prepare Dockerfile from template}
\usage{
prepareDockerfile(study_path, docker_r_version)
}
\description{
Prepare Dockerfile from template
}
\keyword{internal}
