% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{reportValidationResults}
\alias{reportValidationResults}
\title{Report validation results}
\usage{
reportValidationResults(validate_result, verbose = TRUE)
}
\description{
Report validation results
}
\keyword{internal}
