% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewStudyDependencies.R
\name{reviewStudyDependencies}
\alias{reviewStudyDependencies}
\title{Summarise dependencies in renv lock file}
\usage{
reviewStudyDependencies(dir, type = "analysis")
}
\arguments{
\item{dir}{Path to folder containing R project with renv.lock file}

\item{type}{Whether R project is for analysis code or study reporting.}
}
\value{
Summary of state of dependencies
}
\description{
Summarise dependencies in renv lock file
}
