% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGitHub.R
\name{setupGitRemote}
\alias{setupGitRemote}
\title{Setup git remote and push}
\usage{
setupGitRemote(directory, clone_url, user_info = NULL, created_repo = FALSE)
}
\description{
Setup git remote and push
}
\keyword{internal}
