% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{snapshotPackages}
\alias{snapshotPackages}
\title{Snapshot package versions to renv.lock}
\usage{
snapshotPackages(study_path)
}
\arguments{
\item{study_path}{Path to the study directory}
}
\value{
Invisibly returns TRUE
}
\description{
Snapshot package versions to renv.lock
}
