% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{syncBuild}
\alias{syncBuild}
\title{Build Docker image conditionally}
\usage{
syncBuild(study_path, image_name, force_rebuild, packages_changed)
}
\description{
Build Docker image conditionally
}
\keyword{internal}
