% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{syncValidate}
\alias{syncValidate}
\title{Validate Docker environment with rebuild option}
\usage{
syncValidate(study_path, image_name)
}
\description{
Validate Docker environment with rebuild option
}
\keyword{internal}
