% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{verifyImageExists}
\alias{verifyImageExists}
\title{Verify Docker image exists and provide helpful error}
\usage{
verifyImageExists(image_name, check_rstudio = FALSE)
}
\arguments{
\item{image_name}{Name of Docker image}

\item{check_rstudio}{If TRUE, also verify RStudio Server is present}
}
\value{
TRUE (invisibly) if checks pass, otherwise throws error
}
\description{
Verify Docker image exists and provide helpful error
}
\keyword{internal}
