% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{createStudy}
\alias{createStudy}
\title{Creates initial directory for an OMOP CDM network study}
\usage{
createStudy(directory, diagnostics = TRUE, study = TRUE)
}
\arguments{
\item{directory}{Path to a directory that will be used as the root folder
for the study. If it does not exist, it will be created. The directory
must be empty if it already exists.}

\item{diagnostics}{A single TRUE or FALSE value.
If TRUE (the default), the function creates the `diagnostics_code/`
and `diagnostics_shiny/` folders using the package templates.
If FALSE, these diagnostics folders are not created.}

\item{study}{A single TRUE or FALSE value.
If TRUE (the default), the function creates the `study_code/`
and `study_shiny/` folders using the package templates.
If FALSE, these study folders are not created.}
}
\value{
Project directory will be created
}
\description{
Creates initial directory for an OMOP CDM network study
}
\examples{
# Create a study called "SampleStudy" in the current project
study_root <- here::here("SampleStudy")
createStudy(study_root)
list.files(study_root)

# Clean up after running the example
unlink(study_root, recursive = TRUE)
}
