% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{createStudy}
\alias{createStudy}
\title{Create an empty study}
\usage{
createStudy(directory, title = "my_study")
}
\arguments{
\item{directory}{Path to an existing but empty directory which we will be
the root directory for our study}

\item{title}{Study title}
}
\value{
Adds study readme to the
}
\description{
Create an empty study
}
