% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{createStudy}
\alias{createStudy}
\title{Creates initial directory for an OMOP CDM network study}
\usage{
createStudy(directory, diagnostics = TRUE, study = TRUE)
}
\arguments{
\item{directory}{Path to an existing but empty directory which we will be
the root directory for our study}

\item{diagnostics}{If TRUE, directories for diagnositcs will be created}

\item{study}{If TRUE, directories for the study will be created.}
}
\value{
Project directory will be created
}
\description{
Creates initial directory for an OMOP CDM network study
}
