% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{dockerExec}
\alias{dockerExec}
\title{Run a docker command and capture output}
\usage{
dockerExec(args, stream = FALSE)
}
\arguments{
\item{args}{Character vector of docker CLI args}

\item{stream}{If TRUE, stream stdout/stderr to console}
}
\value{
list(ok, status, out, cmd)
}
\description{
Run a docker command and capture output
}
\keyword{internal}
