% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{ensureDocker}
\alias{ensureDocker}
\title{Check if Docker daemon is running}
\usage{
ensureDocker()
}
\value{
TRUE if Docker is available, throws error otherwise
}
\description{
Check if Docker daemon is running
}
\keyword{internal}
