% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopStudyBuilder-package.R
\docType{package}
\name{OmopStudyBuilder-package}
\alias{OmopStudyBuilder}
\alias{OmopStudyBuilder-package}
\title{OmopStudyBuilder: Build Reproducible Network Studies for OMOP Common Data Model}
\description{
Streamlines the setup and execution of network studies using the Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM). Creates standardised project structures with template code, manages dependencies with 'renv', provides code review utilities, and supports containerised execution with 'Docker' for reproducible multi-site studies. Includes 'GitHub' integration for collaboration and version control.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopStudyBuilder/}
  \item Report bugs at \url{https://github.com/oxford-pharmacoepi/OmopStudyBuilder/issues}
}

}
\author{
\strong{Maintainer}: Folu Akintola \email{folu.akintola@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0588-2298}{ORCID})

Authors:
\itemize{
  \item Folu Akintola \email{folu.akintola@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0588-2298}{ORCID})
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
  \item Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

Other contributors:
\itemize{
  \item Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID}) [contributor]
}

}
\keyword{internal}
