% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopStudyBuilder-package.R
\docType{package}
\name{OmopStudyBuilder-package}
\alias{OmopStudyBuilder}
\alias{OmopStudyBuilder-package}
\title{OmopStudyBuilder: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
