% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{runStudy}
\alias{runStudy}
\title{Run OMOP study in Docker container}
\usage{
runStudy(image_name, data_path = NULL, output_path = "./results")
}
\arguments{
\item{image_name}{Name of the Docker image to run}

\item{data_path}{Optional path to data directory (mounted at /data in container)}

\item{output_path}{Where to save results (default: "./results")}
}
\value{
Output path (invisibly)
}
\description{
Run OMOP study in Docker container
}
