% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerWrapper.R
\name{runStudy}
\alias{runStudy}
\title{Run OMOP study in Docker container}
\usage{
runStudy(
  image_name,
  config_file = "config.yml",
  data_path = NULL,
  output_path = "./results"
)
}
\arguments{
\item{image_name}{Name of the Docker image to run}

\item{config_file}{Path to config.yml file (default: "config.yml")}

\item{data_path}{Optional path to data directory to mount}

\item{output_path}{Path where results should be saved (default: "./results")}
}
\value{
Output path (invisibly)
}
\description{
Run OMOP study in Docker container
}
