% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseStudyDependencies.R
\name{summariseStudyDependencies}
\alias{summariseStudyDependencies}
\title{Summarise dependencies in renv lock file}
\usage{
summariseStudyDependencies(dir)
}
\arguments{
\item{dir}{Path to folder containing R project with renv.lock file}
}
\value{
Summary of state of dependencies
}
\description{
Summarise dependencies in renv lock file
}
